+++ duckscript_sdk/src/utils/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod annotation;
 pub(crate) mod condition;
 pub(crate) mod eval;
 pub(crate) mod exec;