+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### v0.6.2
 
-* New function <scope> annotation #121
+* New function &lt;scope&gt; annotation #121
 * New scope_pop_stack command #121
 * New scope_push_stack command #121